refactor: PreparedStatementGenerator 삭제
